Importing Bootstrap JS
Our website's HTML reaches bootstrap's scripts by entering this folder. 
In order to avoid creating a useless temporary file just for the sake of creating this folder, I name this file after one of the scripts that will be uploaded into this folder - so it will be simply overwritten by the actual script.