Importing Bootstrap
Our website's HTML reaches bootstrap's stylesheets by entering this folder. 
In order to avoid creating a useless temporary file just for the sake of creating this folder, I name this file after one of the stylesheets that will be uploaded into this folder - so it will be simply overwritten by the actual stylesheet.